// Downloads/dpm-ui-web-and-mobile-fixes-21-06-2025/dpm-ui-web-and-mobile-fixes/web-apps/consumer-portal/src/types/quoteAndBuy.ts
interface Benefit {
    benefitCategory: string;
    benefitCode: string;
    benefitId: number;
    benefitNameAr: string;
    benefitNameEn: string;
    benefitPrice: number;
    description: string;
    mostPurchased: number;
}